@@ -1,5 +1,7 @@
 using System.Linq;
-using Algorithms.Searches;
+
+using Algorithms.Search;
+
 using NUnit.Framework;
 using NUnit.Framework.Internal;
 
@@ -8,7 +10,7 @@ namespace Algorithms.Tests.Search
     public static class BinarySearcherTests
     {
         [Test]
-        public static void FindIndex_ItemPresent_IndexCorrect([Random(1, 1000, 100)]int n)
+        public static void FindIndex_ItemPresent_IndexCorrect([Random(1, 1000, 100)] int n)
         {
             // Arrange
             var searcher = new BinarySearcher<int>();
@@ -24,7 +26,7 @@ namespace Algorithms.Tests.Search
         }
 
         [Test]
-        public static void FindIndex_ItemMissing_MinusOneReturned([Random(0, 1000, 10)]int n, [Random(-100, 1100, 10)]int missingItem)
+        public static void FindIndex_ItemMissing_MinusOneReturned([Random(0, 1000, 10)] int n, [Random(-100, 1100, 10)] int missingItem)
         {
             // Arrange
             var searcher = new BinarySearcher<int>();
@@ -39,7 +41,7 @@ namespace Algorithms.Tests.Search
         }
 
         [Test]
-        public static void FindIndex_ArrayEmpty_MinusOneReturned([Random(100)]int itemToSearch)
+        public static void FindIndex_ArrayEmpty_MinusOneReturned([Random(100)] int itemToSearch)
         {
             // Arrange
             var searcher = new BinarySearcher<int>();
