@@ -1,6 +1,7 @@
 ﻿using System.Linq;
 using System.Numerics;
 using Algorithms.Sequences;
+using FluentAssertions;
 using NUnit.Framework;
 
 namespace Algorithms.Tests.Sequences
@@ -10,9 +11,9 @@ namespace Algorithms.Tests.Sequences
         [Test]
         public void First10ElementsCorrect()
         {
-            var sequence = new FibonacciSequence().Sequence;
-
-            Assert.AreEqual(new BigInteger[] { 0, 1, 1, 2, 3, 5, 8, 13, 21, 34 }, sequence.Take(10));
+            var sequence = new FibonacciSequence().Sequence.Take(10);
+            sequence.SequenceEqual(new BigInteger[] {0, 1, 1, 2, 3, 5, 8, 13, 21, 34})
+                .Should().BeTrue();
         }
     }
 }
