@@ -1,6 +1,7 @@
 ﻿using System.Linq;
 using System.Numerics;
 using Algorithms.Sequences;
+using FluentAssertions;
 using NUnit.Framework;
 
 namespace Algorithms.Tests.Sequences
@@ -10,9 +11,9 @@ namespace Algorithms.Tests.Sequences
         [Test]
         public void First10ElementsCorrect()
         {
-            var sequence = new NaturalSequence().Sequence;
-
-            Assert.AreEqual(new BigInteger[] { 1, 2, 3, 4, 5, 6, 7, 8, 9, 10 }, sequence.Take(10));
+            var sequence = new NaturalSequence().Sequence.Take(10);
+            sequence.SequenceEqual(new BigInteger[] {1, 2, 3, 4, 5, 6, 7, 8, 9, 10})
+                .Should().BeTrue();
         }
     }
 }
