@@ -1,6 +1,7 @@
 ﻿using System.Linq;
 using System.Numerics;
 using Algorithms.Sequences;
+using FluentAssertions;
 using NUnit.Framework;
 
 namespace Algorithms.Tests.Sequences
@@ -10,9 +11,9 @@ namespace Algorithms.Tests.Sequences
         [Test]
         public void First10ElementsCorrect()
         {
-            var sequence = new PrimesSequence().Sequence;
-
-            Assert.AreEqual(new BigInteger[] { 2, 3, 5, 7, 11, 13, 17, 19, 23, 29 }, sequence.Take(10));
+            var sequence = new PrimesSequence().Sequence.Take(10);
+            sequence.SequenceEqual(new BigInteger[] {2, 3, 5, 7, 11, 13, 17, 19, 23, 29})
+                .Should().BeTrue();
         }
     }
 }
