@@ -1,16 +1,7 @@
 ﻿namespace Algorithms.Numeric.GreatestCommonDivisor
 {
-    /// <summary>
-    /// TODO.
-    /// </summary>
     public interface IGreatestCommonDivisorFinder
     {
-        /// <summary>
-        /// Finds greatest common divisor for numbers a and b.
-        /// </summary>
-        /// <param name="a">TODO.</param>
-        /// <param name="b">TODO. 2.</param>
-        /// <returns>Greatest common divisor.</returns>
-        int Find(int a, int b);
+        int FindGcd(int a, int b);
     }
 }
