@@ -1,4 +1,4 @@
-﻿namespace AStar
+﻿namespace Algorithms.Search.AStar
 {
     /// <summary>
     /// The states the nodes can have.
@@ -8,16 +8,16 @@
         /// <summary>
         /// TODO.
         /// </summary>
-        UNCONSIDERED = 0,
+        Unconsidered = 0,
 
         /// <summary>
         /// TODO.
         /// </summary>
-        OPEN = 1,
+        Open = 1,
 
         /// <summary>
         /// TODO.
         /// </summary>
-        CLOSED = 2,
+        Closed = 2,
     }
 }
