@@ -24,15 +24,15 @@ namespace Algorithms.Sorters.Comparison
 
             // Start from the complete array and one by one
             // reduce current size by one
-            for (var curr_size = n; curr_size > 1; --curr_size)
+            for (var currSize = n; currSize > 1; --currSize)
             {
                 // Find index of the maximum element in
                 // array[0..curr_size-1]
-                var mi = FindMax(array, curr_size, comparer);
+                var mi = FindMax(array, currSize, comparer);
 
                 // Move the maximum element to end of current array
                 // if it's not already at  the end
-                if (mi != curr_size - 1)
+                if (mi != currSize - 1)
                 {
                     // To move to the end, first move maximum
                     // number to beginning
@@ -40,7 +40,7 @@ namespace Algorithms.Sorters.Comparison
 
                     // Now move the maximum number to end by
                     // reversing current array
-                    Flip(array, curr_size - 1);
+                    Flip(array, currSize - 1);
                 }
             }
         }
