@@ -4,7 +4,7 @@ namespace Utilities.Extensions
 {
     public static class DictionaryExtensions
     {
-        public static void AddMany<TKey, TValue>(this Dictionary<TKey, TValue> keys, IEnumerable<(TKey, TValue)> enumerable)
+        public static void AddMany<TKey, TValue>(this Dictionary<TKey, TValue> keys, IEnumerable<(TKey, TValue)> enumerable) where TKey : notnull
         {
             foreach (var (key, value) in enumerable)
             {
